refactor(badge): change folder name from `badges` to `badge`
