+++ src/components.ts
@@ -1,5 +1,5 @@
 export { default as PBackToTopButton } from './navigation/scrolls/back-to-top-button/PBackToTopButton.vue';
-export { default as PBadge } from './data-display/badges/PBadge.vue';
+export { default as PBadge } from './data-display/badge/PBadge.vue';
 export { default as PButton } from './inputs/buttons/button/PButton.vue';
 export { default as PTextButton } from './inputs/buttons/text-button/PTextButton.vue';
 export { default as PEmpty } from './data-display/empty/PEmpty.vue';
