+++ src/data-display/badge/PBadge.stories.mdx
@@ -1,9 +1,9 @@
 import {Meta, Canvas, Story, ArgsTable} from '@storybook/addon-docs/blocks';
-import PBadge from "@/data-display/badges/PBadge";
-import { BADGE_STYLE, BADGE_STYLE_OUTLINE } from '@/data-display/badges/type';
-import { getBadgesArgTypes } from '@/data-display/badges/story-helper';
+import PBadge from "@/data-display/badge/PBadge";
+import { BADGE_STYLE, BADGE_STYLE_OUTLINE } from '@/data-display/badge/type';
+import { getBadgesArgTypes } from '@/data-display/badge/story-helper';
 
-<Meta title='Data Display/Badges' parameters={{
+<Meta title='Data Display/Badge' parameters={{
     design: {
         type: 'figma',
         url: 'https://www.figma.com/file/wq4wSowBcADBuUrMEZLz6i/SpaceONE-Console-Design?node-id=6132%3A124077',
