+++ src/data-display/badge/story-helper.ts
@@ -1,6 +1,6 @@
 import type { ArgTypes } from '@storybook/addons';
 
-import { BADGE_SHAPE, BADGE_STYLE } from '@/data-display/badges/type';
+import { BADGE_SHAPE, BADGE_STYLE } from '@/data-display/badge/type';
 
 
 export const getBadgesArgTypes = (): ArgTypes => ({
