+++ src/data-display/cards/list-card/PListCard.stories.mdx
@@ -7,7 +7,7 @@ import dayjs from 'dayjs';
 import PListCard from './PListCard.vue';
 import PI from '@/foundation/icons/PI.vue';
 import PButton from '@/inputs/buttons/button/PButton.vue';
-import PBadge from '@/data-display/badges/PBadge';
+import PBadge from '@/data-display/badge/PBadge';
 import { CARD_STYLE_TYPE } from '@/data-display/cards/card/config';
 import { getListCardArgTypes } from '@/data-display/cards/list-card/story-helper';
