+++ src/inputs/input/text-input/PTextInput.vue
@@ -125,7 +125,7 @@ import { useFocus } from '@vueuse/core';
 import { unionBy } from 'lodash';
 import type { TranslateResult } from 'vue-i18n';
 
-import PBadge from '@/data-display/badges/PBadge.vue';
+import PBadge from '@/data-display/badge/PBadge.vue';
 import PTag from '@/data-display/tags/PTag.vue';
 import PI from '@/foundation/icons/PI.vue';
 import { useIgnoreWindowArrowKeydownEvents } from '@/hooks';
