+++ src/navigation/tabs/tab/PTab.stories.mdx
@@ -4,7 +4,7 @@ import { reactive, getCurrentInstance, toRefs } from 'vue';
 import { makeOptionalProxy } from '@/utils/composition-helpers';
 import { getTabArgTypes, Inner } from '@/navigation/tabs/tab/story-helper';
 
-import PBadge from '@/data-display/badges/PBadge';
+import PBadge from '@/data-display/badge/PBadge';
 import PTab from './PTab.vue';
